Add stil-id support when registering
